Splitting utils.py into different files
