+++ humancompatible/detect/helpers/utils.py
@@ -249,142 +249,6 @@ def signed_subgroup_prevalence_diff(
     return np.mean(subgroup_b) - np.mean(subgroup_a)
 
 
-def subgroup_map_from_conjuncts_binarized(
-    conjuncts: List[int], X: np.ndarray[np.bool_]
-) -> np.ndarray[np.bool_]:
-    """
-    Generates a boolean subgroup mapping based on the conjunction (AND) of specified features.
-
-    This function creates a boolean array where each element is `True` only if the
-    corresponding row in `X` has `True` values across all columns specified in `conjuncts`.
-    Essentially, it identifies individuals who meet all criteria defined by the conjuncts.
-
-    Args:
-        conjuncts (List[int]): A list of integer indices (column indices) from the
-                               input array `X`. Each index represents a feature
-                               that must be `True` for an individual to be included
-                               in the subgroup.
-        X (np.ndarray[np.bool_]): A 2D NumPy array of boolean values, where rows
-                                  represent individuals and columns represent features.
-
-    Returns:
-        np.ndarray[np.bool_]: A 1D boolean NumPy array (`mapping`) of the same
-                              length as the number of rows in `X`. An element
-                              `mapping[i]` is `True` if `X[i, conj]` is `True` for
-                              all `conj` in `conjuncts`, and `False` otherwise.
-
-    Raises:
-        IndexError: If any index in `conjuncts` is out of bounds for the columns of `X`.
-
-    Examples:
-        >>> import numpy as np
-        >>> X_data = np.array([
-        ...     [True,  True,  False, True],   # Row 0
-        ...     [True,  False, True,  True],   # Row 1
-        ...     [False, True,  True,  False],  # Row 2
-        ...     [True,  True,  True,  True]    # Row 3
-        ... ])
-        >>>
-        >>> # Subgroup where feature at index 0 AND feature at index 1 are True
-        >>> conjuncts_1 = [0, 1]
-        >>> subgroup_map_from_conjuncts_binarized(conjuncts_1, X_data)
-        array([ True, False, False,  True])
-        >>> # Explanation: Only Row 0 and Row 3 have both X[:,0] and X[:,1] as True.
-
-        >>> # Subgroup where feature at index 2 is True
-        >>> conjuncts_2 = [2]
-        >>> subgroup_map_from_conjuncts_binarized(conjuncts_2, X_data)
-        array([False,  True,  True,  True])
-
-        >>> # Subgroup where feature at index 0 AND feature at index 2 are True
-        >>> conjuncts_3 = [0, 2]
-        >>> subgroup_map_from_conjuncts_binarized(conjuncts_3, X_data)
-        array([False,  True, False,  True])
-
-        >>> # Test with an empty list of conjuncts (should return all True)
-        >>> subgroup_map_from_conjuncts_binarized([], X_data)
-        array([ True,  True,  True,  True])
-
-        >>> # Test with an invalid conjunct index (will raise IndexError)
-        >>> try:
-        ...     subgroup_map_from_conjuncts_binarized([0, 99], X_data)
-        ... except IndexError as e:
-        ...     print(e)
-        index 99 is out of bounds for axis 1 with size 4
-    """
-    # Initialize the mapping with all True values. This ensures that if conjuncts
-    # is empty, all individuals are included (logical AND of no conditions is True).
-    mapping = np.ones((X.shape[0],), dtype=bool)
-
-    # Iterate through each specified conjunct (feature index)
-    for conj in conjuncts:
-        # Perform a logical AND operation between the current mapping and the
-        # specified feature column. This filters down the subgroup.
-        mapping &= X[:, conj]  # This will raise IndexError if `conj` is out of bounds
-    return mapping
-
-
-def subgroup_map_from_conjuncts_dataframe(
-    rule: List[Tuple[int, Any]], X: pd.DataFrame
-) -> np.ndarray[np.bool_]:
-    """
-    Build a boolean mask for an MSD rule over a pandas DataFrame.
-
-    Each (index, Bin) in *rule* comes from `detect_bias` or
-    `detect_bias_two_samples`.  We ignore the positional index and
-    use the Bin's `.feature.name`, so this is robust to column re-ordering.
-
-    Args:
-        rule (List[Tuple[int, Any]]): The rule identifying the subgroup, 
-                                      as returned by `detect_bias(...)`.
-        X (pd.DataFrame): The original (protected-only) DataFrame passed 
-                          to `detect_bias`. Must contain all columns named 
-                          in the rule's Bins.
-
-    Returns:
-        np.ndarray[np.bool_]: A 1-D boolean array where True marks rows 
-                              belonging to the subgroup.
-
-    Raises:
-        KeyError: If `X` is missing a column required by the rule.
-    """
-    mask = np.ones(len(X), dtype=bool)
-    for _idx, binop in rule:
-        feat = binop.feature.name
-        if feat not in X.columns:
-            raise KeyError(f"Column '{feat}' required by rule is missing.")
-        col_values = X[feat].to_numpy()
-        mask &= binop.evaluate(col_values)
-    return mask
-
-
-def subgroup_gap(
-    rule: Sequence[Tuple[int, Any]],
-    X: pd.DataFrame,
-    y: pd.Series | np.ndarray,
-    *,
-    signed: bool = True,
-) -> float:
-    """Compute the subgroup discrepancy `delta` or `|delta|` for a given rule.
-
-    Args:
-        rule: Rule returned by `detect_bias` - list of (col_idx, Bin).
-        X: DataFrame containing the protected columns referenced in `rule`.
-        y: Binary outcome vector aligned with `X` (1 = positive outcome).
-        signed: If True returns signed `delta`, else returns `|delta|`.
-
-    Returns:
-        float: Subgroup discrepancy (signed or absolute).
-
-    Raises:
-        KeyError: If `X` is missing a column required by the rule.
-        ValueError: If `y` contains only positives or only negatives.
-    """
-    mask = subgroup_map_from_conjuncts_dataframe(rule, X)
-    fn = signed_subgroup_discrepancy if signed else evaluate_subgroup_discrepancy
-    return float(fn(mask, np.asarray(y).ravel()))
-
-
 def report_subgroup_bias(
     label: str,
     msd: float,
@@ -417,53 +281,3 @@ def report_subgroup_bias(
         # TODO this "=" is not robust to other bins - e.g. continuous ones.
         pretty.append(f"{human_feat} = {human_val}")
     print("Explained rule: " + " AND ".join(pretty))
-
-
-def lin_prog_feas(
-    hist1: np.ndarray,
-    hist2: np.ndarray,
-    delta: float,
-    num_samples: float = 1.0,
-) -> int:
-    """Specifies a number of samples as a fraction of the total
-    histogram bins and checks whether all the sampled bins satisfy
-    
-    |hist1 - hist2| <= delta.
-
-    Args:
-        hist1 (np.ndarray): 1-D array of histogram bin densities for the full dataset.
-        hist2 (np.ndarray): 1-D array of histogram bin densities for the subgroup.
-        delta (float): Threshold for the absolute difference |hist1 - hist2|.
-        num_samples (float): Fraction of total bins to sample.
-            The function draws int(num_samples * (len(hist1) - 1)) random samples.
-
-    Returns:
-        int: Status code from `scipy.optimize.linprog`. A status of 0 indicates
-             the constraints are feasible (i.e., |hist1 - hist2| <= delta for all
-             sampled bins); other codes signal infeasibility or solver errors.
-    """
-    rand_lst1 = []
-    rand_lst2 = []
-
-    for _ in range(0, int(num_samples * (hist1.shape[0] - 1))):
-        i = randrange(0, hist1.shape[0] - 1)
-        rand_lst1.append(float(hist1[i]))
-        rand_lst2.append(float(hist2[i]))
-
-    rand_arr1 = np.expand_dims(np.array(rand_lst1), axis=1)
-    rand_arr2 = np.expand_dims(np.array(rand_lst2), axis=1)
-
-    # We are not interested in the optimization itself, but in the
-    # feasibility of the problem, therefore the coefficient in the
-    # objective function is set to 0 and the only variable (x_0) is
-    # fixed at 1
-    c = 0
-    x0_bounds = (1, 1)
-
-    # Accomodate for the + & - signs of the absolute value in
-    # |r_a1 - r_a2| <= delta
-    A_ub = np.vstack((rand_arr1, -rand_arr1))
-    b_ub = np.vstack((delta + rand_arr2, delta - rand_arr2))
-
-    res = optimize.linprog(c, A_ub, b_ub, bounds=[x0_bounds])
-    return res.status
